winedbg: Adjust the parameters of the backend read/write routines to match those of ReadProcessMemory/WriteProcessMemory, since those are the ones actually used.

+++ programs/winedbg/be_alpha.c
@@ -87,7 +87,7 @@ static unsigned be_alpha_insert_Xpoint(HANDLE hProcess, const struct be_process_
                                        void* addr, unsigned long* val, unsigned size)
 {
     unsigned long       xbp;
-    unsigned long       sz;
+    SIZE_T              sz;
 
     switch (type)
     {
